InputDemoをファイル分割
