+++ input/gamepad/Input.cpp
@@ -15,7 +15,7 @@ Input::Input() {
         m_inputs.insert(std::make_pair(id, PlayerInput(id)));
     }
     // TOdO:
-    m_sources.insert(std::make_pair(GamePadId::_1P, InputSource::JoyConHorizontally));
+    m_sources.insert(std::make_pair(GamePadId::_1P, InputSource::JoyConFrontally));
     m_sources.insert(std::make_pair(GamePadId::_2P, InputSource::JoyConHorizontally));
     m_sources.insert(std::make_pair(GamePadId::_3P, InputSource::JoyConHorizontally));
     m_sources.insert(std::make_pair(GamePadId::_4P, InputSource::JoyConHorizontally));
